+++ Sources/MapboxSearch/InternalAPI/CoreAliases.swift
@@ -10,6 +10,7 @@ typealias CoreSearchOptions = MapboxCoreSearch.SearchOptions
 typealias CoreBoundingBox = MapboxCoreSearch.LonLatBBox
 typealias CoreSearchResult = MapboxCoreSearch.SearchResult
 typealias CoreAttributeSet = MapboxCoreSearch.AttributeSet
+typealias CoreRetrieveOptions = MapboxCoreSearch.RetrieveOptions
 typealias CoreRoutablePoint = MapboxCoreSearch.RoutablePoint
 typealias CoreResultMetadata = MapboxCoreSearch.ResultMetadata
 typealias CoreRequestOptions = MapboxCoreSearch.RequestOptions