+++ Sources/MapboxSearch/PublicAPI/IndexableDataResolver.swift
@@ -8,4 +8,15 @@ public protocol IndexableDataResolver {
     ///   - suggestion: suggestion to resolve
     ///   - completion: completion closure
     func resolve(suggestion: SearchResultSuggestion, completion: @escaping (SearchResult?) -> Void)
+
+    /// Resolves ``SearchResultSuggestion`` into ``SearchResult`` if possible
+    /// - Parameters:
+    ///   - suggestion: suggestion to resolve
+    ///   - retrieveOptions: Define attribute sets to request additional metadata attributes
+    ///   - completion: completion closure
+    func resolve(
+        suggestion: SearchResultSuggestion,
+        retrieveOptions: RetrieveOptions?,
+        completion: @escaping (SearchResult?) -> Void
+    )
 }