+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 - [SearchOptions] Remove SearchOptions.AttributeSets parameter.
 
+- [SearchResult] Add ResultChildMetadata type to SearchResultMetadata.children field.
+
 **MapboxCommon**: v24.6.0
 **MapboxCoreSearch**: v2.3.0-rc.6